Add SQLFederationEngine interface to standardize sql federation query engine calls
