+++ kernel/sql-federation/core/src/test/java/org/apache/shardingsphere/sqlfederation/engine/StandardSQLFederationEngineTest.java
@@ -32,6 +32,7 @@ import org.apache.shardingsphere.sql.parser.api.CacheOption;
 import org.apache.shardingsphere.sqlfederation.config.SQLFederationRuleConfiguration;
 import org.apache.shardingsphere.sqlfederation.engine.fixture.rule.SQLFederationDeciderRuleMatchFixture;
 import org.apache.shardingsphere.sqlfederation.engine.fixture.rule.SQLFederationDeciderRuleNotMatchFixture;
+import org.apache.shardingsphere.sqlfederation.engine.impl.StandardSQLFederationEngine;
 import org.apache.shardingsphere.sqlfederation.rule.SQLFederationRule;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -51,7 +52,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class SQLFederationEngineTest {
+class StandardSQLFederationEngineTest {
     
     private final DatabaseType databaseType = TypedSPILoader.getService(DatabaseType.class, "MySQL");
     
@@ -70,15 +71,15 @@ class SQLFederationEngineTest {
         when(queryContext.getSqlStatementContext()).thenReturn(sqlStatementContext);
         ShardingSphereDatabase database = new ShardingSphereDatabase("foo_db", databaseType, mock(ResourceMetaData.class, RETURNS_DEEP_STUBS), mock(RuleMetaData.class), Collections.emptyList());
         when(queryContext.getUsedDatabase()).thenReturn(database);
-        SQLFederationEngine engine = createSQLFederationEngine(globalRules, Collections.emptyList());
+        StandardSQLFederationEngine engine = createSQLFederationEngine(globalRules, Collections.emptyList());
         assertTrue(engine.decide(queryContext, mock(RuleMetaData.class)));
         engine.close();
     }
     
-    private SQLFederationEngine createSQLFederationEngine(final Collection<ShardingSphereRule> globalRules, final Collection<ShardingSphereRule> databaseRules) {
+    private StandardSQLFederationEngine createSQLFederationEngine(final Collection<ShardingSphereRule> globalRules, final Collection<ShardingSphereRule> databaseRules) {
         when(metaData.getDatabase("foo_db").getRuleMetaData().getRules()).thenReturn(databaseRules);
         when(metaData.getGlobalRuleMetaData()).thenReturn(new RuleMetaData(globalRules));
-        return new SQLFederationEngine("foo_db", "foo_db", metaData, mock(ShardingSphereStatistics.class), mock(JDBCExecutor.class));
+        return new StandardSQLFederationEngine("foo_db", "foo_db", metaData, mock(ShardingSphereStatistics.class), mock(JDBCExecutor.class));
     }
     
     @Test
@@ -86,7 +87,7 @@ class SQLFederationEngineTest {
         Collection<ShardingSphereRule> globalRules =
                 Collections
                         .singletonList(new SQLFederationRule(new SQLFederationRuleConfiguration(false, false, mock(CacheOption.class)), Collections.emptyList()));
-        SQLFederationEngine engine = createSQLFederationEngine(globalRules, Collections.emptyList());
+        StandardSQLFederationEngine engine = createSQLFederationEngine(globalRules, Collections.emptyList());
         RuleMetaData globalRuleMetaData = new RuleMetaData(globalRules);
         assertFalse(engine.decide(mock(QueryContext.class), globalRuleMetaData));
         engine.close();
@@ -104,7 +105,7 @@ class SQLFederationEngineTest {
         QueryContext queryContext = mock(QueryContext.class);
         when(queryContext.getSqlStatementContext()).thenReturn(selectStatementContext);
         when(queryContext.getUsedDatabase()).thenReturn(database);
-        SQLFederationEngine engine = createSQLFederationEngine(globalRules, Collections.emptyList());
+        StandardSQLFederationEngine engine = createSQLFederationEngine(globalRules, Collections.emptyList());
         RuleMetaData globalRuleMetaData = new RuleMetaData(globalRules);
         assertTrue(engine.decide(queryContext, globalRuleMetaData));
         engine.close();
@@ -114,7 +115,7 @@ class SQLFederationEngineTest {
     void assertDecideWhenExecuteNotSelectStatement() throws SQLException {
         Collection<ShardingSphereRule> globalRules =
                 Collections.singletonList(new SQLFederationRule(new SQLFederationRuleConfiguration(true, false, mock(CacheOption.class)), Collections.emptyList()));
-        SQLFederationEngine engine = createSQLFederationEngine(globalRules, Collections.emptyList());
+        StandardSQLFederationEngine engine = createSQLFederationEngine(globalRules, Collections.emptyList());
         RuleMetaData globalRuleMetaData = new RuleMetaData(globalRules);
         assertFalse(engine.decide(mock(QueryContext.class), globalRuleMetaData));
         engine.close();
@@ -133,7 +134,7 @@ class SQLFederationEngineTest {
         QueryContext queryContext = mock(QueryContext.class);
         when(queryContext.getSqlStatementContext()).thenReturn(selectStatementContext);
         when(queryContext.getUsedDatabase()).thenReturn(database);
-        SQLFederationEngine engine = createSQLFederationEngine(globalRules, databaseRules);
+        StandardSQLFederationEngine engine = createSQLFederationEngine(globalRules, databaseRules);
         RuleMetaData globalRuleMetaData = new RuleMetaData(globalRules);
         assertTrue(engine.decide(queryContext, globalRuleMetaData));
         engine.close();
@@ -152,7 +153,7 @@ class SQLFederationEngineTest {
         QueryContext queryContext = mock(QueryContext.class);
         when(queryContext.getSqlStatementContext()).thenReturn(selectStatementContext);
         when(queryContext.getUsedDatabase()).thenReturn(database);
-        SQLFederationEngine engine = createSQLFederationEngine(globalRules, databaseRules);
+        StandardSQLFederationEngine engine = createSQLFederationEngine(globalRules, databaseRules);
         assertFalse(engine.decide(queryContext, new RuleMetaData(globalRules)));
         engine.close();
     }
@@ -172,7 +173,7 @@ class SQLFederationEngineTest {
         when(queryContext.getSqlStatementContext()).thenReturn(selectStatementContext);
         when(queryContext.getParameters()).thenReturn(Collections.emptyList());
         when(queryContext.getUsedDatabase()).thenReturn(database);
-        SQLFederationEngine engine = createSQLFederationEngine(globalRules, databaseRules);
+        StandardSQLFederationEngine engine = createSQLFederationEngine(globalRules, databaseRules);
         assertTrue(engine.decide(queryContext, new RuleMetaData(globalRules)));
         engine.close();
     }
